Ch 2: Partially apply viewThumbnail function
All Elm functions are curried. When we partially apply `viewThumbnail`
with just a single argument, we get back a function that accepts the
missing `thumb` argument (defined in `viewThumbnail` function).

So, for each of the `model.photos` going through `List.map`, the
argument gets passed into the partial function that was created.