Ch 3: Render different thumbnail sizes
There is no default branch `_` in our case-expression here as we defined
that we have the type annotation that `size` (argument) is a
`ThumbnailSize` type. It can only have the three values defined in
`ThumbnailSize` type.

Elm's compiler will know if you covered every possibility and will not
compile if you forget one of the cases.