Ch 3: Convert `Msg` to custom type
Custom type allows us to implement all the necessary cases using
case-expression. We should try to limit using `_ ->` case and let the
missing-patterns error from the compiler to help us catch for cases that
we forgot.

By using custom type, we can hold only the minimum amount of data
required for each variant instead of holding data that might not be
required in another variant.