Ch 3: Randomly pick photo with command + Random.generate
`Random.generate` returns a command that generates random values wrapped
in messages. It takes two arguments: A `Random.Generator` and a function
that can wrap the resulting value.

`Random.generate` will pass the `Msg` `GotSelectedIndex` and let our
`update` function handle the message. In our case, we update our model
in the case-expression.