Ch 4: Decode JSON response from server
Making use of `NoRedInk/elm-json-decode-pipeline` instead of
`Json.Decode.mapX` in order to look for required or optional fields in a
JSON response from the server.

We use `Http.expectJson` and pass in a `List Photo` decoder and
simplified the `Ok` branch of our `GotPhotos` message. Since the `Ok`
would be a `List Photo` type, we can just use `(::)` in order to get the
first element of the list and the remainder.