Ch 5: Event from JS updates model in Elm
Created a custom event called "slide" which will contain a JavaScript
object dispatched with this event. In Elm, we use `Html.Events.on` to
decode the "slide" event, coupled with `Json.Decode.map` to convert the
`Json.Decode.at` traversal of our JavaScript object into our message.

Instead of creating concise code that will allow us to add more filters
into a list of filters (inside our Model), we chose to go with more
verbosity to prevent typos. This allows us to make the tradeoff of
having concise code to catch bugs before they appear in production.
Refer to pages 139-140 of the book.