Ch 5: Fix initial photo load with filters
The problem we had previously when the first photo didn't load was due
to the timing of the DOM updates. The `<canvas>` element wasn't loaded
and `Pasta.apply` was silently failing.

To solve that, we delay our call to `Pasta.apply` in our JavaScript code
by wrapping it with a `requestAnimationFrame`. This allows us to run the
`Pasta.apply` call _after the next time_ `view` gets an updated DOM.